クリティカルレートの値を入力 その他の細かいミスの訂正

+++ project1/Enemy.h
@@ -16,7 +16,7 @@ public:
 	int Gold;
 	int EXP;
 	Element Element;
-	int CriticalRate;
+	float CriticalRate;
 
 	Enemy(const EnemyData& data)
 		: ID(data.ID),
